+++ CLAUDE.md
@@ -88,7 +88,7 @@ Run a single test class/method using absl conventions, e.g. `python tests/test_o
 - `registry.py` — global `_OPTIMIZERS` dict with a `@register_optimizer("name")` decorator; `get_optimizer_cls` / `get_configured_optimizer_cls` provide lookup. **Important: a class is only in the registry once its module has been imported.** External callers must import the optimizer module (e.g. `from emerging_optimizers.orthogonalized_optimizers import muon`) before `get_optimizer_cls("muon")` will find it.
 - `mixin.py` — `WeightDecayMixin` with four `weight_decay_method` modes: `decoupled`, `independent`, `l2`, `palm`. Optimizers that include this mixin set `self.weight_decay_method` and call `_apply_weight_decay_inplace(p, grad, lr, wd)` inside `step`.
 - `orthogonalized_optimizers/` — Muon and friends (`muon`, `adaptive_muon`, `muon_hyperball`, `mop`, `polargrad`, `scion`, `sinkhorn_muon`, `spel`). All extend `OrthogonalizedOptimizer` (`orthogonalized_optimizer.py`), which implements the EMA-momentum + (optional Nesterov) + `scaled_orthogonalize_fn` pipeline. Subclasses customize behavior by overriding `orthogonalize(p, grad, **group_kwargs)` (e.g. for split-QKV preconditioning) or the `pre_weight_update_fn_inplace` / `post_weight_update_fn_inplace` hooks (used by hyperball-style updates that preserve weight norms). `muon_utils` holds the Newton–Schulz iteration coefficients and helpers, including the tensor-parallel `newton_schulz_tp`; `spectral_clipping_utils` implements spectral norm clipping.
-- `soap/` — the SOAP family, sharing `soap_utils.py`: `soap.py` (`SOAP`, plus `StackedSoap` which optimizes batched/3D params through transient 2D stacking), `rekls.py` (REKLS), `tp_rekls.py` (experimental tensor-parallel REKLS), and `moso.py` (MOSO, momentum one-sided SOAP). Uses Kronecker-factored preconditioning and tracks eigenbases per-parameter.
+- `soap/` — the SOAP family, sharing `soap_utils.py`: `soap.py` (`SOAP`), `stacked_soap.py` (`StackedSoap`, which optimizes batched/3D params through transient 2D stacking), `rekls.py` (REKLS), `tp_rekls.py` (experimental tensor-parallel REKLS), and `moso.py` (MOSO, momentum one-sided SOAP). Uses Kronecker-factored preconditioning and tracks eigenbases per-parameter.
 - `psgd/` — `PSGDPro` (`psgd.py`, registered as `psgd_pro`) plus `psgd_kron_contractions.py` (Einsum contractions for triangular factors), `procrustes_step.py`, and `psgd_utils.py`.
 - `scalar_optimizers/` — element-wise optimizers (`lion`, `signum`, `laprop`, `sim_ademamix`), defined in `__init__.py` on top of the `base.py` hierarchy (`SingleMomentumOptimizer` / `TwoMomentsOptimizer`); `update_functions/` holds the reusable per-algorithm update primitives (adam, lion, laprop, ademamix, madam, signum).
 - `riemannian_optimizers/` — `normalized_optimizer.py` (manifold-normalized updates; `oblique_sgd`, `oblique_adam`).
@@ -129,5 +129,5 @@ CI runs in containers built via `docker/Dockerfile.ci`. Key env flags set by the
 
 ## Rules for agent
 
-* Never write or modify comment including docstring unless explicitly asked.
+* Never write new comments or docstrings unless explicitly asked. You may fix existing comments or docstrings that are stale or incorrect (e.g. references to renamed, removed, or moved code, or a description that no longer matches the code) so they stay accurate.
 * Never run tests unless explicitly asked.
